requirements.txt
fastapi
uvicorn
langgraph
langchain-ollama
langchain-openai
langchain-anthropic
tavily-python
python-dotenv
fpdf
chromadb
pydantic
tiktoken
